Changed detached toolbar to dynamically resize based on the number of icons in the toolbar and their size.

+++ src/main/java/com/redmondsims/gistfx/preferences/settings/Get.java
@@ -177,8 +177,8 @@ public class Get {
 		return prefs.getDouble(LABEL.DIVIDER_EXPANDED.Name(), 0.0);
 	}
 
-	public double iconBaseSize() {
-		return prefs.getDouble(LABEL.ICON_BASE_SIZE.Name(), 25.0);
+	public Integer iconBaseSize() {
+		return prefs.getInt(LABEL.ICON_BASE_SIZE.Name(), 25);
 	}
 
 	public ColorOption trayIconColorOption() {
